+++ Sources/NablaCore/Data/Authentication/AuthTokens.swift
@@ -1,6 +1,6 @@
 import Foundation
 
-public struct Tokens {
+public struct AuthTokens {
     public let accessToken: String
     public let refreshToken: String